# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API de Gerenciamento de Produtos",
    "version": "0.1.0"
  },
  "paths": {
    "/produtos": {
      "get": {
        "summary": "Listar Produtos",
        "operationId": "listar_produtos_produtos_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProdutoRead"
                  },
                  "type": "array",
                  "title": "Response Listar Produtos Produtos Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar Produto",
        "operationId": "criar_produto_produtos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProdutoCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/produtos/{produto_id}": {
      "get": {
        "summary": "Buscar Produto",
        "operationId": "buscar_produto_produtos__produto_id__get",
        "parameters": [
          {
            "name": "produto_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Produto Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Deletar Produto",
        "operationId": "deletar_produto_produtos__produto_id__delete",
        "parameters": [
          {
            "name": "produto_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Produto Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ProdutoCreate": {
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 1,
            "title": "Nome"
          },
          "preco": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Preco"
          },
          "estoque": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Estoque",
            "default": 0
          },
          "ativo": {
            "type": "boolean",
            "title": "Ativo",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "nome",
          "preco"
        ],
        "title": "ProdutoCreate",
        "example": {
          "ativo": true,
          "estoque": 10,
          "nome": "Teclado Mecânico",
          "preco": 350.0
        }
      },
      "ProdutoRead": {
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 1,
            "title": "Nome"
          },
          "preco": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Preco"
          },
          "estoque": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Estoque",
            "default": 0
          },
          "ativo": {
            "type": "boolean",
            "title": "Ativo",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "nome",
          "preco",
          "id"
        ],
        "title": "ProdutoRead",
        "example": {
          "ativo": true,
          "estoque": 5,
          "id": 1,
          "nome": "Notebook Gamer",
          "preco": 4999.9
        }
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}